backend/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
pydantic
pydantic-settings
openai
websockets
python-socketio
aiohttp
email-validator
